pyproject.toml
[project]
name = "Battling Package"
version = "1.0.0"
description = "Battling package, made by @wascertified"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "discord.py>=2.5.0",
]

[tool.dexi]
public = true
ballsdex-version = ">=2.22.0"
include-license = true

[tool.dexi.package]
source = "battle"
target = "battling"